- Finally created wiringpi-disabling config flag. That should ease debugging/development a bit. - Cut the current index into template base and created GPIO monitor template. - Re-organized static files a bit and added a few CDN files to it, since app has a high chance to work disconnected from internet. - Minor restructurization
