+++ config.py
@@ -14,6 +14,7 @@ WEB_SERVICE_CONFIG = {
     "host": "0.0.0.0",
     "port": 5000,
     "debug": False,
+    "use_wiringpi": False,
     "security_secret_key": "super-secret",
     "security_password_hash": "bcrypt",
     "security_password_salt": "$2a$06$6sSyl34Zw.48NBXwGBSURO",
