--- utils/web.py
@@ -1,6 +1,0 @@
-def get_formatter():
-    """
-    Logger format provider
-    :return: Format String
-    """
-    return "%(asctime)s; %(levelname)s; %(message)s"
