+++ web/routes/general.py
@@ -2,6 +2,7 @@
 General routing library
 """
 
+from flask import render_template
 from flask_security import login_required, roles_required, roles_accepted
 
 from web import web_service
@@ -13,6 +14,6 @@ from web import web_service
 def monitor():
     """
     Maps monitor path to a static index file
-    :return: Static file
+    :return: Rendered template
     """
-    return web_service.send_static_file("index.html")
+    return render_template("general/gpio.html")
